Introduce NumVersionsToKeep (#474)
Allow users to specify how many versions to store per key. By default, this is set to one, because most users are only interested in the latest version of the key.

Fix: Instead of using the readTs by Oracle, we should be using the minimum readTs of any pending transaction. We should never discard any versions > minReadTs, only versions which are <= minReadTs.