Set readTs in write batch to latest done readTs so that we can discard older versions of keys during compactions.
Every Transaction stores the latest value of `readTs` it is aware of. When the transaction is discarded (which happens even when we commit), the global value of `readMark` is updated.
https://github.com/dgraph-io/badger/blob/ef05d3439792607477618c0164d9a6e977f43a63/txn.go#L501-L503
Previously, the `readTs` of transaction inside the write batch struct was set to 0. So the global value of `readMark` would also be set to 0 (unless someone ran a transaction after using write batch).
Due to the 0 value of the global `readMark`, the compaction algorithm would skip all the values which were inserted in the write batch call.
See https://github.com/dgraph-io/badger/blob/ef05d3439792607477618c0164d9a6e977f43a63/levels.go#L480-L484
and
https://github.com/dgraph-io/badger/blob/ef05d3439792607477618c0164d9a6e977f43a63/txn.go#L138-L145
The `o.readMark.DoneUntil()` call would always return `0` and so the compaction wouldn't compact the newer values.

With this commit, the compaction algorithm works as expected with key-values inserted via Transaction API or via the Write Batch API.

See https://github.com/dgraph-io/badger/issues/767