Set move key's expiresAt for keys with TTL (#1006)
The value log GC moves keys from one value log file to another. The
current implementation of vlog GC would not copy the TTL from the
original entry to the new move key. This commit fixes that.

Without this change, the move keys were not being removed as seen
in issue https://github.com/dgraph-io/badger/issues/974.
With this commit, the move keys are being removed.