Rework Block and Index cache (#1473)
This PR separates the block cache and the index cache. Earlier we had a mix of both the caches. 
Block Cache -> Stores block. Should be used when running badger with compression/encryption
Index Cache -> Used to limit the memory used by table indices and bloom filters. 

This PR also remove the `KeepBlocksInCache` and `KeepBlockIndicesInCache` options.