fix(Graphql): Ensure GraphQL hasInverse works (#34)
In the past, hasInverse in GraphQL had been avoiding the use of reverse directive available at the core. After a careful review, I decided it's best to remove the `@reverse` directive from the core. It was adding a bunch of complexity, which was unwarranted. The same can be achieved by adding edges in both the directions at the GraphQL layer.

In this PR, I've added support for hasInverse back after the rewrite of https://github.com/outcaste-io/outserv/pull/32 ( 9b9d56 ).

BREAKING: This change removes `@reverse` directive from Outserv core.

Fixes #35 .