feat(Upsert): Upsert from within Raft loop (#42)
This PR adds the ability to query for and determine the UID for a node, based on its XIDs (external IDs) within the Raft loop. It does so by figuring out which predicates constitute the XID, and adding those in a new `NewObjects` struct within Mutations.

This would ensure that concurrent upserts do not end up creating duplicate nodes. Raft loop serializes all the writes. Whichever upsert comes first would create the node. The second upsert would then be able to query the created node and reuse it, without failing. This is a better mechanism than doing optimistic transactions which would have resulted in aborts.

This PR also consolidates a bunch of things into one:
- Removes Posting.ValType.
- Removes DirectedEdge.
- Converts NQuad to Edge.
- Removes message Value previously used by NQuad.

Essentially, it consolidates the type conversion and mutation structs, introduced in 2016, into simplified structures. Hence eliminating a lot of roundabout complexity gained over time.