fix(GraphQL): Ensure that a list of scalars remain sorted (#76)
If we have a list of strings, currently Outserv stores them sorted by their fingerprints. Instead, as per the GraphQL spec, we need to maintain their order. This PR achieves that outcome with a new ListType. The list values are binary encoded and stored together as one value. This ensures correct ordering.

This means a posting list can only have a maximum of one posting for values (UIDs are stored in bitmaps anyway).

This PR also removes the `type` func from DQL. Instead, now we use `has` func for resolving queries with no filters. More work is needed to clean up `dgraph.type` predicate.

chunker interface is now replaced with jsonChunker. That's the only chunker we have after RDF chunker was removed.