feat(Boot): Use type information in boot loader (#80)
The type information can be generated by either passing in the type of the root object as a flag or by having a @type field in the objects. Once a type for the parent object is determined, the children object types are deduced from the GraphQL Schema.

All this info is used to generate UIDs for the objects if none is provided.